Moving the 'home' package into the package 'ui'
