+++ app/src/main/java/com/arctouch/codechallenge/injection/module/ActivityBuildersModule.kt
@@ -1,7 +1,7 @@
 package com.arctouch.codechallenge.injection.module
 
 import com.arctouch.codechallenge.base.BaseActivity
-import com.arctouch.codechallenge.home.HomeActivity
+import com.arctouch.codechallenge.ui.home.HomeActivity
 import com.arctouch.codechallenge.splash.SplashActivity
 import com.arctouch.codechallenge.ui.movieDetail.MovieDetailActivity
 import dagger.Module
