+++ app/src/main/java/com/arctouch/codechallenge/injection/module/ViewModelModule.kt
@@ -2,7 +2,7 @@ package com.arctouch.codechallenge.injection.module
 
 import android.arch.lifecycle.ViewModel
 import android.arch.lifecycle.ViewModelProvider
-import com.arctouch.codechallenge.home.HomeViewModel
+import com.arctouch.codechallenge.ui.home.HomeViewModel
 import com.arctouch.codechallenge.injection.ViewModelFactory
 import com.arctouch.codechallenge.injection.ViewModelKey
 import com.arctouch.codechallenge.ui.movieDetail.MovieDetailViewModel
