+++ app/src/main/java/com/arctouch/codechallenge/splash/SplashActivity.kt
@@ -6,8 +6,7 @@ import com.arctouch.codechallenge.R
 import com.arctouch.codechallenge.api.TmdbApi
 import com.arctouch.codechallenge.base.BaseActivity
 import com.arctouch.codechallenge.data.Cache
-import com.arctouch.codechallenge.home.HomeActivity
-import com.arctouch.codechallenge.ui.movieDetail.MovieDetailActivity
+import com.arctouch.codechallenge.ui.home.HomeActivity
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
 
@@ -22,7 +21,7 @@ class SplashActivity : BaseActivity() {
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe {
                 Cache.cacheGenres(it.genres)
-                startActivity(Intent(this, MovieDetailActivity::class.java))
+                startActivity(Intent(this, HomeActivity::class.java))
                 finish()
             }
     }
